apps/vanna-ai/requirements.txt
fastapi
uvicorn
psycopg2-binary # PostgreSQL driver
SQLAlchemy
python-dotenv
groq # Or other LLM libraries you choose
vanna